gcp - move ConfigRegionsTest to test_query.py
config_regions() lives in query.py; its tests belong alongside it rather
than in test_region.py, which covers Region itself.
