+++ tools/c7n_gcp/tests/test_query.py
@@ -1,11 +1,46 @@
 # Copyright The Cloud Custodian Authors.
 # SPDX-License-Identifier: Apache-2.0
 
+import os
+
+from c7n.config import Config
 from c7n.resources import load_resources
-from c7n_gcp.query import GcpLocation
+from c7n_gcp.query import config_regions, GcpLocation
 from c7n_gcp.provider import GoogleCloud
 
 from gcp_common import BaseTest
+
+
+class ConfigRegionsTest(BaseTest):
+
+    def test_no_explicit_region(self):
+        # real gcp cli run with no --region: commands.py seeds
+        # options.region = "" and, unlike aws, the gcp provider never
+        # expands per-policy regions (GoogleCloud.initialize_policies is a
+        # no-op), so config.region stays "".
+        self.assertEqual(config_regions(Config.empty(region="")), ())
+
+    def test_region_aws_default_sentinel(self):
+        # Config.empty()'s own default region must not be treated as an
+        # explicit gcp region, whether or not AWS_DEFAULT_REGION is set.
+        default_region = os.environ.get('AWS_DEFAULT_REGION', 'us-east-1')
+        self.assertEqual(config_regions(Config.empty(region=default_region)), ())
+
+    def test_explicit_region(self):
+        self.assertEqual(
+            config_regions(Config.empty(region="us-central1")), ("us-central1",))
+
+    def test_explicit_regions(self):
+        self.assertEqual(
+            config_regions(Config.empty(regions=("us-central1", "us-east1"))),
+            ("us-central1", "us-east1"))
+
+    def test_regions_all(self):
+        self.assertEqual(config_regions(Config.empty(regions=("all",))), ())
+
+    def test_regions_all_overrides_other_entries(self):
+        self.assertEqual(
+            config_regions(Config.empty(regions=("us-central1", "all"))), ())
 
 
 def test_gcp_resource_metadata_asset_type():
